fixed rendering of cubes
Fixed cube batch renderering, need to fix batch rendering colors and textures, finish abstracting stuff that is leaking into main.cpp

+++ glfwGlewTest/src/Camera.cpp
@@ -36,7 +36,7 @@ void Camera::Matrix(float FOVdeg, float nearPlane, float farPlane, Shader& shade
 	// Makes camera look in the right direction from the right position
 	view = glm::lookAt(Position, Position + Orientation, Up);
 	// Adds perspective to the scene
-	projection = glm::perspective(glm::radians(FOVdeg), (float)width / height, nearPlane, farPlane);
+	projection = glm::perspective(glm::radians(FOVdeg), (float)width / (float)height, nearPlane, farPlane);
 
 	// Exports the camera matrix to the Vertex Shader
 	shader.SetUniformMat4f(uniform, (projection * view));
